pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantinvest"
version = "1.0.0"
description = "Factor-based portfolio construction: risk models, optimization, and backtesting"
authors = [
  { name = "Zhang Shizhe", email = "shizhe_zhang@berkeley.edu" },
]
readme = "quantinvest/README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
keywords = ["portfolio", "optimization", "risk", "finance", "quantitative"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Financial and Insurance Industry",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Office/Business :: Financial :: Investment",
  "Topic :: Scientific/Engineering :: Mathematics",
]

# Kept minimal and aligned to actual imports under quantinvest/
dependencies = [
  "numpy>=1.21.0",
  "pandas>=1.5.0",
  "scipy>=1.9.0",
  "scikit-learn>=1.1.0",
  "cvxpy>=1.2.0",
  "matplotlib>=3.5.0",
  "seaborn>=0.11.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=3.0.0",
  "black>=22.0.0",
  "flake8>=4.0.0",
  "mypy>=0.991",
  "isort>=5.10.0",
]
data = [
  "pandas-datareader>=0.10.0",
  "plotly>=5.0.0",
]
solvers = [
  "mosek>=9.0.0",
  "gurobi>=9.0.0",
]

[project.urls]
Homepage = "https://github.com/minikouda/QuantitativeInvestment_ZJU"
Repository = "https://github.com/minikouda/QuantitativeInvestment_ZJU"
Documentation = "https://github.com/minikouda/QuantitativeInvestment_ZJU/blob/main/quantinvest/README.md"
Issues = "https://github.com/minikouda/QuantitativeInvestment_ZJU/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["quantinvest*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\\.pyi?$'
extend-exclude = """
/(
  \\.eggs
  | \\.git
  | \\.hg
  | \\.mypy_cache
  | \\.tox
  | \\.venv
  | build
  | dist
)/
"""

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
# If you later add real tests under tests/, pytest will pick them up.
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "--cov=quantinvest",
  "--cov-report=html",
  "--cov-report=term-missing",
  "--strict-markers",
  "--strict-config",
  "--verbose",
]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests",
  "unit: marks tests as unit tests",
]
